+++ nall/arguments.hpp
@@ -1,8 +1,10 @@
 #pragma once
 
+#include <nall/string.hpp>
+#include <nall/directory.hpp>
+#include <nall/file.hpp>
 #include <nall/location.hpp>
 #include <nall/path.hpp>
-#include <nall/string.hpp>
 #include <nall/vector.hpp>
 
 namespace nall {
