+++ nall/directory.hpp
@@ -192,83 +192,6 @@ inline auto directory::copy(const string& source, const string& target) -> bool 
     }
     return _wrmdir(utf16_t(pathname)) == 0;
   }
-
-  inline auto directory::exists(const string& pathname) -> bool {
-    if(!pathname) return false;
-    string name = pathname;
-    name.trim("\"", "\"");
-    DWORD result = GetFileAttributes(utf16_t(name));
-    if(result == INVALID_FILE_ATTRIBUTES) return false;
-    return (result & FILE_ATTRIBUTE_DIRECTORY);
-  }
-
-  inline auto directory::ufolders(const string& pathname, const string& pattern) -> vector<string> {
-    if(!pathname) {
-      //special root pseudo-folder (return list of drives)
-      wchar_t drives[PATH_MAX] = {0};
-      GetLogicalDriveStrings(PATH_MAX, drives);
-      wchar_t* p = drives;
-      while(*p || *(p + 1)) {
-        if(!*p) *p = ';';
-        p++;
-      }
-      return string{(const char*)utf8_t(drives)}.replace("\\", "/").split(";");
-    }
-
-    vector<string> list;
-    string path = pathname;
-    path.transform("/", "\\");
-    if(!path.endsWith("\\")) path.append("\\");
-    path.append("*");
-    HANDLE handle;
-    WIN32_FIND_DATA data;
-    handle = FindFirstFile(utf16_t(path), &data);
-    if(handle != INVALID_HANDLE_VALUE) {
-      if(wcscmp(data.cFileName, L".") && wcscmp(data.cFileName, L"..")) {
-        if(data.dwFileAttributes & FILE_ATTRIBUTE_DIRECTORY) {
-          string name = (const char*)utf8_t(data.cFileName);
-          if(name.match(pattern)) list.append(name);
-        }
-      }
-      while(FindNextFile(handle, &data) != false) {
-        if(wcscmp(data.cFileName, L".") && wcscmp(data.cFileName, L"..")) {
-          if(data.dwFileAttributes & FILE_ATTRIBUTE_DIRECTORY) {
-            string name = (const char*)utf8_t(data.cFileName);
-            if(name.match(pattern)) list.append(name);
-          }
-        }
-      }
-      FindClose(handle);
-    }
-    return list;
-  }
-
-  inline auto directory::ufiles(const string& pathname, const string& pattern) -> vector<string> {
-    if(!pathname) return {};
-
-    vector<string> list;
-    string path = pathname;
-    path.transform("/", "\\");
-    if(!path.endsWith("\\")) path.append("\\");
-    path.append("*");
-    HANDLE handle;
-    WIN32_FIND_DATA data;
-    handle = FindFirstFile(utf16_t(path), &data);
-    if(handle != INVALID_HANDLE_VALUE) {
-      if((data.dwFileAttributes & FILE_ATTRIBUTE_DIRECTORY) == 0) {
-        string name = (const char*)utf8_t(data.cFileName);
-        if(name.match(pattern)) list.append(name);
-      }
-      while(FindNextFile(handle, &data) != false) {
-        if((data.dwFileAttributes & FILE_ATTRIBUTE_DIRECTORY) == 0) {
-          string name = (const char*)utf8_t(data.cFileName);
-          if(name.match(pattern)) list.append(name);
-        }
-      }
-      FindClose(handle);
-    }
-    return list;
-  }
 #else
   inline auto directoryIsFolder(DIR* dp, struct dirent* ep) -> bool {
     if(ep->d_type == DT_DIR) return true;
@@ -353,3 +276,7 @@ inline auto directory::copy(const string& source, const string& target) -> bool 
 #endif
 
 }
+
+#if defined(NALL_HEADER_ONLY)
+  #include <nall/directory.cpp>
+#endif
