+++ nall/http/response.hpp
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <nall/chrono.hpp>
+#include <nall/file-map.hpp>
 #include <nall/http/message.hpp>
 
 namespace nall::HTTP {
