+++ nall/inode.hpp
@@ -31,15 +31,7 @@ struct inode {
     return access(name, X_OK) == 0;
   }
 
-  static auto hidden(const string& name) -> bool {
-    #if defined(PLATFORM_WINDOWS)
-    auto attributes = GetFileAttributes(utf16_t(name));
-    return attributes & FILE_ATTRIBUTE_HIDDEN;
-    #else
-    //todo: is this really the best way to do this? stat doesn't have S_ISHIDDEN ...
-    return name.split("/").last().beginsWith(".");
-    #endif
-  }
+  static auto hidden(const string& name) -> bool;
 
   static auto mode(const string& name) -> u32 {
     struct stat data{};
@@ -161,3 +153,7 @@ struct inode {
 };
 
 }
+
+#if defined(NALL_HEADER_ONLY)
+  #include <nall/inode.cpp>
+#endif
