+++ nall/memory.hpp
@@ -195,47 +195,11 @@ template<u32 size, typename T> auto writem(void* target, T data) -> void {
   for(s32 n = size - 1; n >= 0; n--) *p++ = data >> n * 8;
 }
 
-inline auto map(u32 size, bool executable) -> void* {
-  #if defined(API_WINDOWS)
-  DWORD protect = executable ? PAGE_EXECUTE_READWRITE : PAGE_READWRITE;
-  return VirtualAlloc(nullptr, size, MEM_RESERVE | MEM_COMMIT, protect);
-  #elif defined(API_POSIX)
-  int prot = PROT_READ | PROT_WRITE;
-  int flags = MAP_ANON | MAP_PRIVATE;
-  if(executable) {
-    prot |= PROT_EXEC;
-    #if defined(PLATFORM_MACOS)
-    flags |= MAP_JIT;
-    #endif
-  }
-  return mmap(nullptr, size, prot, flags, -1, 0);
-  #else
-  return nullptr;
-  #endif
-}
+auto map(u32 size, bool executable) -> void*;
 
-inline auto unmap(void* target, u32 size) -> void {
-  #if defined(API_WINDOWS)
-  VirtualFree(target, 0, MEM_RELEASE);
-  #elif defined(API_POSIX)
-  munmap(target, size);
-  #endif
-}
+auto unmap(void* target, u32 size) -> void;
 
-inline auto protect(void* target, u32 size, bool executable) -> void {
-  #if defined(API_WINDOWS)
-  DWORD protect = executable ? PAGE_EXECUTE_READWRITE : PAGE_READWRITE;
-  DWORD oldProtect;
-  VirtualProtect(target, size, protect, &oldProtect);
-  #elif defined(API_POSIX)
-  int prot = PROT_READ | PROT_WRITE;
-  if(executable) {
-    prot |= PROT_EXEC;
-  }
-  int ret = mprotect(target, size, prot);
-  assert(ret == 0);
-  #endif
-}
+auto protect(void* target, u32 size, bool executable) -> void;
 
 inline auto jitprotect(bool executable) -> void {
   #if defined(PLATFORM_MACOS)
@@ -246,3 +210,7 @@ inline auto jitprotect(bool executable) -> void {
 }
 
 }
+
+#if defined(NALL_HEADER_ONLY)
+  #include <nall/memory.cpp>
+#endif
