+++ nall/path.hpp
@@ -24,73 +24,19 @@ inline auto real(string_view name) -> string {
   return result;
 }
 
-inline auto program() -> string {
-  #if defined(PLATFORM_WINDOWS)
-  wchar_t path[PATH_MAX] = L"";
-  GetModuleFileName(nullptr, path, PATH_MAX);
-  string result = (const char*)utf8_t(path);
-  result.transform("\\", "/");
-  return Path::real(result);
-  #else
-  Dl_info info;
-  dladdr((void*)&program, &info);
-  return Path::real(info.dli_fname);
-  #endif
-}
+auto program() -> string;
 
 // /
 // c:/
-inline auto root() -> string {
-  #if defined(PLATFORM_WINDOWS)
-  wchar_t path[PATH_MAX] = L"";
-  SHGetFolderPathW(nullptr, CSIDL_WINDOWS | CSIDL_FLAG_CREATE, nullptr, 0, path);
-  string result = (const char*)utf8_t(path);
-  result.transform("\\", "/");
-  return slice(result, 0, 3);
-  #else
-  return "/";
-  #endif
-}
+auto root() -> string;
 
 // /home/username/
 // c:/users/username/
-inline auto user() -> string {
-  #if defined(PLATFORM_WINDOWS)
-  wchar_t path[PATH_MAX] = L"";
-  SHGetFolderPathW(nullptr, CSIDL_PROFILE | CSIDL_FLAG_CREATE, nullptr, 0, path);
-  string result = (const char*)utf8_t(path);
-  result.transform("\\", "/");
-  #else
-  struct passwd* userinfo = getpwuid(getuid());
-  string result = userinfo->pw_dir;
-  #endif
-  if(!result) result = ".";
-  if(!result.endsWith("/")) result.append("/");
-  return result;
-}
+auto user() -> string;
 
 // /home/username/Desktop/
 // c:/users/username/Desktop/
-inline auto desktop(string_view name = {}) -> string {
-  #if defined(PLATFORM_WINDOWS)
-  wchar_t path[PATH_MAX] = L"";
-  SHGetFolderPathW(nullptr, CSIDL_DESKTOP | CSIDL_FLAG_CREATE, nullptr, 0, path);
-  string result = (const char*)utf8_t(path);
-  result.transform("\\", "/");
-  #elif defined(PLATFORM_MACOS)
-  string result = {user(), "Desktop/"};
-  #else
-  string result;
-  if(const char *env = getenv("XDG_DESKTOP_DIR")) {
-    result = string(env);
-  } else {
-    result = {user(), "Desktop/"};
-  }
-  #endif
-  if(!result) result = ".";
-  if(!result.endsWith("/")) result.append("/");
-  return result.append(name);
-}
+auto desktop(string_view name = {}) -> string;
 
 //todo: MacOS uses the same location for userData() and userSettings()
 //... is there a better option here?
@@ -98,85 +44,24 @@ inline auto desktop(string_view name = {}) -> string {
 // /home/username/.config/
 // ~/Library/Application Support/
 // c:/users/username/appdata/roaming/
-inline auto userSettings() -> string {
-  #if defined(PLATFORM_WINDOWS)
-  wchar_t path[PATH_MAX] = L"";
-  SHGetFolderPathW(nullptr, CSIDL_APPDATA | CSIDL_FLAG_CREATE, nullptr, 0, path);
-  string result = (const char*)utf8_t(path);
-  result.transform("\\", "/");
-  #elif defined(PLATFORM_MACOS)
-  string result = {Path::user(), "Library/Application Support/"};
-  #else
-  string result;
-  if(const char *env = getenv("XDG_CONFIG_HOME")) {
-    result = string(env);
-  } else {
-    result = {Path::user(), ".config/"};
-  }
-  #endif
-  if(!result) result = ".";
-  if(!result.endsWith("/")) result.append("/");
-  return result;
-}
+auto userSettings() -> string;
 
 // /home/username/.local/share/
 // ~/Library/Application Support/
 // c:/users/username/appdata/local/
-inline auto userData() -> string {
-  #if defined(PLATFORM_WINDOWS)
-  wchar_t path[PATH_MAX] = L"";
-  SHGetFolderPathW(nullptr, CSIDL_LOCAL_APPDATA | CSIDL_FLAG_CREATE, nullptr, 0, path);
-  string result = (const char*)utf8_t(path);
-  result.transform("\\", "/");
-  #elif defined(PLATFORM_MACOS)
-  string result = {Path::user(), "Library/Application Support/"};
-  #else
-  string result;
-  if(const char* env = getenv("XDG_DATA_HOME")) {
-    result = string(env);
-  } else {
-    result = {Path::user(), ".local/share/"};
-  }
-  #endif
-  if(!result) result = ".";
-  if(!result.endsWith("/")) result.append("/");
-  return result;
-}
+auto userData() -> string;
 
 // /usr/share
 // /Library/Application Support/
 // c:/ProgramData/
-inline auto sharedData() -> string {
-  #if defined(PLATFORM_WINDOWS)
-  wchar_t path[PATH_MAX] = L"";
-  SHGetFolderPathW(nullptr, CSIDL_COMMON_APPDATA | CSIDL_FLAG_CREATE, nullptr, 0, path);
-  string result = (const char*)utf8_t(path);
-  result.transform("\\", "/");
-  #elif defined(PLATFORM_MACOS)
-  string result = "/Library/Application Support/";
-  #else
-  string result = "/usr/share/";
-  #endif
-  if(!result) result = ".";
-  if(!result.endsWith("/")) result.append("/");
-  return result;
-}
+auto sharedData() -> string;
 
 // /tmp
 // c:/users/username/AppData/Local/Temp/
-inline auto temporary() -> string {
-  #if defined(PLATFORM_WINDOWS)
-  wchar_t path[PATH_MAX] = L"";
-  GetTempPathW(PATH_MAX, path);
-  string result = (const char*)utf8_t(path);
-  result.transform("\\", "/");
-  #elif defined(P_tmpdir)
-  string result = P_tmpdir;
-  #else
-  string result = "/tmp/";
-  #endif
-  if(!result.endsWith("/")) result.append("/");
-  return result;
-}
+auto temporary() -> string;
 
 }
+
+#if defined(NALL_HEADER_ONLY)
+  #include <nall/path.cpp>
+#endif
