+++ nall/platform.hpp
@@ -9,18 +9,12 @@ namespace Math {
 }
 
 #if defined(PLATFORM_WINDOWS)
-  #include <nall/windows/guard.hpp>
-  #include <initguid.h>
-  #include <cguid.h>
-  #include <winsock2.h>
-  #include <ws2tcpip.h>
-  #include <windows.h>
+  #if defined(NALL_HEADER_ONLY)
+    #include <nall/windows/windows.hpp>
+  #endif
   #include <direct.h>
   #include <io.h>
   #include <wchar.h>
-  #include <shlobj.h>
-  #include <shellapi.h>
-  #include <nall/windows/guard.hpp>
   #include <nall/windows/utf8.hpp>
 #endif
 
@@ -73,20 +67,20 @@ namespace Math {
 #endif
 
 #if defined(PLATFORM_WINDOWS)
-  #undef  IN
-  #undef  OUT
-  #undef  interface
   #define dllexport __declspec(dllexport)
   #define MSG_NOSIGNAL 0
+  #define PATH_MAX 260
 
-  extern "C" {
-    using pollfd = WSAPOLLFD;
-  }
+  #if !defined(INVALID_HANDLE_VALUE)
+    #define INVALID_HANDLE_VALUE ((HANDLE)-1)
+  #endif
+
+  typedef void* HANDLE;
 
   inline auto access(const char* path, int amode) -> int { return _waccess(nall::utf16_t(path), amode); }
   inline auto getcwd(char* buf, size_t size) -> char* { wchar_t wpath[PATH_MAX] = L""; if(!_wgetcwd(wpath, size)) return nullptr; strcpy(buf, nall::utf8_t(wpath)); return buf; }
   inline auto mkdir(const char* path, int mode) -> int { return _wmkdir(nall::utf16_t(path)); }
-  inline auto poll(struct pollfd fds[], unsigned long nfds, int timeout) -> int { return WSAPoll(fds, nfds, timeout); }
+  inline auto poll(struct pollfd fds[], unsigned long nfds, int timeout) -> int;
   inline auto putenv(const char* value) -> int { return _wputenv(nall::utf16_t(value)); }
   inline auto realpath(const char* file_name, char* resolved_name) -> char* { wchar_t wfile_name[PATH_MAX] = L""; if(!_wfullpath(wfile_name, nall::utf16_t(file_name), PATH_MAX)) return nullptr; strcpy(resolved_name, nall::utf8_t(wfile_name)); return resolved_name; }
   inline auto rename(const char* oldname, const char* newname) -> int { return _wrename(nall::utf16_t(oldname), nall::utf16_t(newname)); }
@@ -94,17 +88,9 @@ namespace Math {
   namespace nall {
     //network functions take void*, not char*. this allows them to be used without casting
 
-    inline auto recv(int socket, void* buffer, size_t length, int flags) -> ssize_t {
-      return ::recv(socket, (char*)buffer, length, flags);
-    }
-
-    inline auto send(int socket, const void* buffer, size_t length, int flags) -> ssize_t {
-      return ::send(socket, (const char*)buffer, length, flags);
-    }
-
-    inline auto setsockopt(int socket, int level, int option_name, const void* option_value, socklen_t option_len) -> int {
-      return ::setsockopt(socket, level, option_name, (const char*)option_value, option_len);
-    }
+    auto recv(int socket, void* buffer, size_t length, int flags) -> ssize_t;
+    auto send(int socket, const void* buffer, size_t length, int flags) -> ssize_t;
+    auto setsockopt(int socket, int level, int option_name, const void* option_value, int option_len) -> int;
   }
 #else
   #define dllexport
@@ -194,3 +180,7 @@ inline auto spinloop() -> void {
 
 #define export $export
 #define register $register
+
+#if defined(NALL_HEADER_ONLY)
+  #include <nall/platform.cpp>
+#endif
