+++ nall/windows/guid.hpp
@@ -4,14 +4,10 @@
 
 namespace nall {
 
-inline auto guid() -> string {
-  GUID guidInstance;
-  CoCreateGuid(&guidInstance);
-
-  wchar_t guidString[39];
-  StringFromGUID2(guidInstance, guidString, 39);
-
-  return (char*)utf8_t(guidString);
-}
+auto guid() -> string;
 
 }
+
+#if defined(NALL_HEADER_ONLY)
+  #include <nall/windows/guid.cpp>
+#endif
