+++ nall/windows/utf8.hpp
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <nall/stdint.hpp>
+
 namespace nall {
   //UTF-8 to UTF-16
   struct utf16_t {
@@ -9,15 +11,7 @@ namespace nall {
     utf16_t(const utf16_t&) = delete;
     auto operator=(const utf16_t&) -> utf16_t& = delete;
 
-    auto operator=(const char* s) -> utf16_t& {
-      reset();
-      if(!s) s = "";
-      length = MultiByteToWideChar(CP_UTF8, 0, s, -1, nullptr, 0);
-      buffer = new wchar_t[length + 1];
-      MultiByteToWideChar(CP_UTF8, 0, s, -1, buffer, length);
-      buffer[length] = 0;
-      return *this;
-    }
+    auto operator=(const char* s) -> utf16_t&;
 
     operator wchar_t*() { return buffer; }
     operator const wchar_t*() const { return buffer; }
@@ -45,15 +39,7 @@ namespace nall {
     utf8_t(const utf8_t&) = delete;
     auto operator=(const utf8_t&) -> utf8_t& = delete;
 
-    auto operator=(const wchar_t* s) -> utf8_t& {
-      reset();
-      if(!s) s = L"";
-      length = WideCharToMultiByte(CP_UTF8, 0, s, -1, nullptr, 0, nullptr, nullptr);
-      buffer = new char[length + 1];
-      WideCharToMultiByte(CP_UTF8, 0, s, -1, buffer, length, nullptr, nullptr);
-      buffer[length] = 0;
-      return *this;
-    }
+    auto operator=(const wchar_t* s) -> utf8_t&;
 
     auto reset() -> void {
       delete[] buffer;
@@ -73,12 +59,10 @@ namespace nall {
     u32 length = 0;
   };
 
-  inline auto utf8_arguments(int& argc, char**& argv) -> void {
-    wchar_t** wargv = CommandLineToArgvW(GetCommandLineW(), &argc);
-    argv = new char*[argc + 1]();
-    for(u32 i = 0; i < argc; i++) {
-      argv[i] = new char[PATH_MAX];
-      strcpy(argv[i], nall::utf8_t(wargv[i]));
-    }
-  }
+  auto utf8_arguments(int& argc, char**& argv) -> void;
+
 }
+
+#if defined(NALL_HEADER_ONLY)
+  #include <nall/windows/utf8.cpp>
+#endif
