+++ ruby/video/direct3d.cpp
@@ -1,7 +1,4 @@
-#undef interface
-#define interface struct
 #include <d3d9.h>
-#undef interface
 
 static LRESULT CALLBACK VideoDirect3D9_WindowProcedure(HWND hwnd, UINT msg, WPARAM wparam, LPARAM lparam) {
   if(msg == WM_SYSKEYDOWN && wparam == VK_F4) return false;
